Split summary storage logic from calculation
