+++ cubedash/summary/__init__.py
@@ -1,3 +1,3 @@
 
-from ._summarise import TimePeriodOverview
+from ._model import TimePeriodOverview
 from ._stores import SummaryStore
