+++ cubedash/summary/_schema.py
@@ -15,7 +15,7 @@ from sqlalchemy import func, Table, Column, ForeignKey, String, \
 from sqlalchemy.dialects import postgresql as postgres
 from sqlalchemy.types import UserDefinedType
 
-from ._summarise import GridCell
+from ._model import GridCell
 
 CUBEDASH_SCHEMA = 'cubedash'
 METADATA = MetaData(schema=CUBEDASH_SCHEMA)
