+++ integration_tests/test_summarise_data.py
@@ -1,3 +1,8 @@
+"""
+Load a lot of real-world DEA datasets (very slow)
+
+And then check their statistics match expected.
+"""
 from typing import Set, Optional
 
 import pytest
@@ -62,9 +67,9 @@ def populate_index(module_dea_index):
 def test_calc_month(summary_store: SummaryStore):
     # One Month
     _expect_values(
-        summary_store.calculate_summary(
+        summary_store.update(
             'ls8_nbar_scene',
-            _utils.as_time_range(2017, 4)
+            2017, 4, None
         ),
         dataset_count=408,
         footprint_count=408,
